MAJOR Satble version without unpacking and state saving * Implemented pause/resume * Fixed displaying file name and speed * Libraries added to repository

+++ gui/progressdialog.h
@@ -11,7 +11,7 @@ public:
 				   HANDLE continue_event);
 	~ProgressDialog();
 	bool SetDisplayedData(const StlString &fname, 
-						  unsigned int speed,
+						  double speed,
 						  unsigned int file_progress, 
 						  unsigned int total_progress);
 	bool Create();
